Fixed some authentication issues

+++ backend/controllers/user.js
@@ -3,15 +3,20 @@ import express from 'express';
 import User from "../models/user.models.js";
 import Post from '../models/post.models.js';
 import Section from '../models/section.models.js';
-import { getUsername } from '../utils/utils.js';
+import { getUsername, checkIfAuthenticated } from '../utils/utils.js';
+import { generateAccessToken } from '../middlewares/auth.js';
 
 const router = express.Router();
 
 export const profileRoute = async (req, res) => {
   try{
-    User.findOne({username: getUsername(req, res)}, (err, user) => {
-      res.status(200).json(user);
-    })
+    if(checkIfAuthenticated === true){
+      User.findOne({username: getUsername(req, res)}, (err, user) => {
+        res.status(200).json(user);
+      })
+    }else{
+      res.status(202).json({"message": "Not authenticated"});
+    }
   }catch{
     res.status(500).json({"message": "Error"});
   }
@@ -19,13 +24,17 @@ export const profileRoute = async (req, res) => {
 
 export const updateProfileRoute = async (req, res) => {
   try{
-    const {email, password} = req.body;
-    User.findOne({username: getUsername(req, res)}, (err, user) => {
-      user.email = email || user.email;
-      user.password = password || user.email;
-      user.save();
-      res.status(200).json({"message": "success"})
-    })
+    if(checkIfAuthenticated === true){
+      const {email, password} = req.body;
+      User.findOne({username: getUsername(req, res)}, (err, user) => {
+        user.email = email || user.email;
+        user.password = password || user.email;
+        user.save();
+        res.status(200).json({"message": "success"})
+      })
+    }else{
+      res.status(202).json({"message": "Not authenticated"});
+    }
   }catch{
     res.status(500).json({"message": "Error"})
   }
@@ -33,12 +42,16 @@ export const updateProfileRoute = async (req, res) => {
 
 export const deleteProfileRoute = async (req, res) => {
   try{
-    const username = getUsername(req, res);
-    await User.deleteOne({username: username});
-    await Post.deleteMany({user: username});
-    await Section.deleteMany({user: username});
-
-    res.status(200).json({"message": "success"});
+    if(checkIfAuthenticated === true){
+      const username = getUsername(req, res);
+      await User.deleteOne({username: username});
+      await Post.deleteMany({user: username});
+      await Section.deleteMany({user: username});
+      
+      res.status(200).json({"message": "success"});
+    }else{
+      res.status(202).json({"message": "Not authenticated"})
+    }
   }catch{
     res.status(500).json({"message": "Error"});
   }
